third_party/WebKit/LayoutTests/inspector/console/console-filter-level-test-expected.txt

CONSOLE INFO: line 10: sample info
CONSOLE MESSAGE: line 11: sample log
CONSOLE WARNING: line 12: sample warning
CONSOLE DEBUG: line 13: sample debug
CONSOLE ERROR: line 14: sample error
CONSOLE INFO: line 16: abc info
CONSOLE INFO: line 17: def info
CONSOLE WARNING: line 19: abc warn
CONSOLE WARNING: line 20: def warn
Tests that console can filter messages by source.


Running: beforeFilter
beforeFilter
>console-filter-level-test.html:10 sample info
>console-filter-level-test.html:11 sample log
>console-filter-level-test.html:12 sample warning
>console-filter-level-test.html:13 sample debug
>console-filter-level-test.html:14 sample error
>console-filter-level-test.html:16 abc info
>console-filter-level-test.html:17 def info
>console-filter-level-test.html:19 abc warn
>console-filter-level-test.html:20 def warn
>'Should be always visible'
>"Should be always visible"

Running: onlyWarning
>console-filter-level-test.html:12 sample warning
>console-filter-level-test.html:19 abc warn
>console-filter-level-test.html:20 def warn
>'Should be always visible'
>"Should be always visible"

Running: onlyLog
>console-filter-level-test.html:11 sample log
>'Should be always visible'
>"Should be always visible"

Running: onlyErrorDebug
>console-filter-level-test.html:13 sample debug
>console-filter-level-test.html:14 sample error
>'Should be always visible'
>"Should be always visible"

Running: onlyAbcMessagePlain
>console-filter-level-test.html:16 abc info
>console-filter-level-test.html:19 abc warn
>'Should be always visible'
>"Should be always visible"

Running: onlyAbcMessageRegex
>console-filter-level-test.html:16 abc info
>console-filter-level-test.html:19 abc warn
>'Should be always visible'
>"Should be always visible"

Running: onlyAbcMessageRegexWarning
>console-filter-level-test.html:19 abc warn
>'Should be always visible'
>"Should be always visible"
